requirements.txt
Django==1.8.6
Markdown==2.6.4
MySQL-python==1.2.5
PyYAML==3.11
SQLAlchemy==0.8.0b2
appnope==0.1.0
decorator==4.0.4
django-autocomplete-light==2.2.10
django-chosen==0.1
django-filter==0.11.0
django-grappelli==2.7.2
django-remote-forms==0.0.1
django-rest-swagger==0.3.4
django-smart-selects==1.1.1
djangorestframework==3.3.1
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
wsgiref==0.1.2
